+++ src/internals.rs
@@ -1,0 +1,7 @@
+//! Expose various unstable EtherCrab internals.
+//!
+//! Anything exported by this module should be considered unstable and may change at any time.
+
+pub use crate::eeprom::device_reader::DeviceEeprom;
+pub use crate::eeprom::EepromDataProvider;
+pub use crate::slave::slave_client::SlaveClient;